+++ packages/frontend/src/hooks/useOrganizationRoles.ts
@@ -3,9 +3,10 @@ import {
     type Role,
     type RoleAssignment,
 } from '@lightdash/common';
-import { useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { lightdashApi } from '../api';
 import useApp from '../providers/App/useApp';
+import useToaster from './toaster/useToaster';
 import useQueryError from './useQueryError';
 
 export const useOrganizationRoles = (loadScopes?: boolean) => {
@@ -53,6 +54,44 @@ export const useOrganizationRoleAssignments = () => {
         {
             enabled: !!user.data?.organizationUuid,
             onError: (result: ApiError) => setErrorResponse(result),
+        },
+    );
+};
+
+export const useUpsertOrganizationUserRoleAssignmentMutation = () => {
+    const { user } = useApp();
+    const queryClient = useQueryClient();
+    const { showToastSuccess, showToastApiError } = useToaster();
+
+    return useMutation(
+        async ({ userId, roleId }: { userId: string; roleId: string }) => {
+            if (!user.data?.organizationUuid) {
+                throw new Error('Organization UUID is required');
+            }
+            return lightdashApi({
+                url: `/orgs/${user.data.organizationUuid}/roles/assignments/user/${userId}`,
+                method: 'POST',
+                body: JSON.stringify({ roleId }),
+                version: 'v2',
+            });
+        },
+        {
+            onSuccess: async () => {
+                await queryClient.invalidateQueries([
+                    'organization_role_assignments',
+                ]);
+                await queryClient.invalidateQueries(['organization_users']);
+                await queryClient.refetchQueries(['user']);
+                showToastSuccess({
+                    title: 'Success! User role updated.',
+                });
+            },
+            onError: ({ error }: { error: any }) => {
+                showToastApiError({
+                    title: "Failed to update user's role",
+                    apiError: error,
+                });
+            },
         },
     );
 };